media-source/SourceBuffer-timestampOffset_removed.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Setting SourceBuffer#timestampOffset after the sourceBuffer object is removed</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
var mimes = ['video/webm; codecs="vorbis,vp8"', 'video/mp4'];

//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function testMedia(mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e) {
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            mediaSource.removeSourceBuffer(sourceBuffer);
            t.step(function() {
                assert_throws({name: 'InvalidStateError'},
                              function() {
                                  sourceBuffer.timestampOffset = 1;
                              },
                              'Check for setting timestampOffset of removed sourceBuffer object.');
            });
            t.done();
        });
        video.src = window.URL.createObjectURL(mediaSource);
    }, 'Setting SourceBuffer#timestampOffset (' + mime + '): ' +
       'If sourceBuffer object has been removed from the sourceBuffers attribute of the parent media source, ' +
       'then INVALID_STATE_ERR exception should be thrown.');
}
mimes.forEach(function(mime) {
    testMedia(mime);
});
</script>
</body>
</html>
